Move CoreData serialisation into coredata.cpp
