+++ src/classes/CMakeLists.txt
@@ -38,7 +38,6 @@ add_library(
   isotopologueset.cpp
   isotopologueweight.cpp
   kvector.cpp
-  masters.cpp
   molecule.cpp
   moleculedistributor.cpp
   neutronweights.cpp
